+++ crates/cli-tools/test.sh
@@ -19,4 +19,5 @@ set -e
 
 test_helper
 
+cargo test --lib --features=action
 cargo test --lib