+++ crates/protocol-usb/src/lib.rs
@@ -19,6 +19,10 @@
 #![no_std]
 #![feature(never_type)]
 
+use wasefire_one_of::exactly_one_of;
+
+exactly_one_of!["device", "host"];
+
 extern crate alloc;
 #[cfg(feature = "std")]
 extern crate std;