+++ crates/runner-host/src/board.rs
@@ -12,16 +12,16 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+pub mod applet;
 pub mod button;
 mod crypto;
 mod debug;
 mod led;
-mod platform;
+pub mod platform;
 mod rng;
 mod storage;
 pub mod timer;
 pub mod uart;
-#[cfg(feature = "usb")]
 pub mod usb;
 
 use tokio::sync::mpsc::Sender;
@@ -37,8 +37,8 @@ pub struct State {
     pub led: bool,
     pub timers: timer::Timers,
     pub uarts: uart::Uarts,
-    #[cfg(feature = "usb")]
-    pub usb: usb::Usb,
+    pub protocol: platform::protocol::State,
+    pub usb: usb::State,
     pub storage: Option<FileStorage>,
     #[cfg(feature = "web")]
     pub web: web_server::Client,
@@ -63,6 +63,7 @@ impl Api for Board {
         }
     }
 
+    type Applet = applet::Impl;
     type Button = button::Impl;
     type Crypto = crypto::Impl;
     type Debug = debug::Impl;
@@ -72,6 +73,5 @@ impl Api for Board {
     type Storage = storage::Impl;
     type Timer = timer::Impl;
     type Uart = uart::Impl;
-    #[cfg(feature = "usb")]
     type Usb = usb::Impl;
 }