+++ crates/prelude/CHANGELOG.md
@@ -1,9 +1,19 @@
 # Changelog
 
-## 0.6.1-git
+## 0.7.0-git
+
+### Major
+
+- Remove `debug::{assert,assert_eq}()` in favor of `core::{assert,assert_eq}!`
+- Remove `debug::exit()` in favor of `scheduling::{abort,exit}()`
+
+### Minor
+
+- Add `scheduling::exit()` for successful applet exit
 
 ### Patch
 
+- Make sure at compile-time that at most one `native`, `test`, or `wasm` feature is enabled
 - Update dependencies
 
 ## 0.6.0
@@ -162,4 +172,4 @@
 
 ## 0.1.0
 
-<!-- Increment to skip CHANGELOG.md test: 1 -->
+<!-- Increment to skip CHANGELOG.md test: 2 -->