+++ crates/scheduler/CHANGELOG.md
@@ -1,9 +1,22 @@
 # Changelog
 
-## 0.3.2-git
+## 0.4.0-git
+
+### Major
+
+- Remove `board-api-platform{,-protocol,-update}` features
+
+### Minor
+
+- Support `PlatformLock` protocol call
+- Support `AppletExitStatus` protocol call (the platform keeps running when the applet exits)
+- Support `Applet{Install,Uninstall}` protocol calls
+- Migrate from `debug::exit()` to `scheduling::exit()`
+- Support `PlatformUpdate{Metadata,Transfer}` protocol calls
 
 ### Patch
 
+- Make sure at compile-time that exactly one `native` or `wasm` feature is enabled
 - Use `derive-where` instead of `derivative`
 - Implement `defmt::Format` for `Key` when `defmt` is enabled
 - Stop using `log::Debug2Format()` when logging events
@@ -152,4 +165,4 @@
 
 ## 0.1.0
 
-<!-- Increment to skip CHANGELOG.md test: 4 -->
+<!-- Increment to skip CHANGELOG.md test: 9 -->